+++ lib/flutter_bluesky.dart
@@ -218,7 +218,8 @@ class FlutterBluesky extends Bluesky {
   Future<List<Map>?> upload(List<ImageFile> imgFiles) async {
     List<Map>? images = [];
     for (var imgFile in imgFiles) {
-      Tuple2 res = await plugin.uploadBlob(imgFile.bytes, imgFile.mimeType!);
+      Tuple2 res =
+          await plugin.uploadBlob(await imgFile.bytes, imgFile.mimeType!);
       images.add({"image": res.item2["blob"], "alt": ""});
     }
     if (images.isEmpty) {