+++ lib/util/image_util.dart
@@ -1,10 +1,12 @@
 import 'dart:io';
-import 'dart:typed_data';
-
-import 'package:file_picker/file_picker.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:path/path.dart' as path;
-import 'package:image/image.dart' as img;
+import 'package:image_picker/image_picker.dart';
+
+const imageQuality = 70;
+const double maxWidth = 1440;
+const ImageSource source = ImageSource.gallery;
 
 /// This util deals with file_picker and image_picker.
 class ImageUtil {
@@ -28,41 +30,43 @@ class ImageUtil {
     "webp": "image/webp",
   };
 
-  static Image image(PlatformFile file) {
-    return file.bytes != null
-        ? Image.memory(file.bytes!) // web case
-        : Image.file(File(file.path!)); // ios, android case
+  static Future<List<XFile>> pickMultiImage() async {
+    return await ImagePicker().pickMultiImage(
+      imageQuality: imageQuality,
+      maxWidth: maxWidth,
+    );
+  }
+
+  static Future<XFile?> pickImage() async {
+    return await ImagePicker().pickImage(
+      imageQuality: imageQuality,
+      maxWidth: maxWidth,
+      source: source,
+    );
+  }
+
+  static Image image(XFile pickedFile) {
+    if (kIsWeb) {
+      // see https://pub.dev/packages/image_picker_for_web#limitations-on-the-web-platform
+      return Image.network(pickedFile.path); // web case
+    } else {
+      return Image.file(File(pickedFile.path)); // ios, android case
+    }
   }
 }
 
+// https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/embed/images.json
+const int maxSize = 1000000; // 1MB
+
 class ImageFile {
-  static const maxWidth = 450;
-  final PlatformFile file;
-  final Uint8List origin;
+  final XFile file;
   final String? mimeType;
+
   ImageFile(this.file)
-      : origin = file.bytes ?? File(file.path!).readAsBytesSync(),
-        mimeType = ImageUtil.exts[path.extension(file.name).substring(1)];
+      : mimeType = ImageUtil.exts[path.extension(file.name).substring(1)];
 
-  Uint8List get bytes {
-    img.Image? original = img.decodeImage(origin);
-    if (original!.width > maxWidth) {
-      img.Image resized = img.copyResize(original, width: maxWidth);
-      // debugPrint("resized!.width: ${resized.width}");
-      return encode(resized);
-    } else {
-      return origin;
-    }
-  }
-
-  Uint8List encode(img.Image image) {
-    switch (mimeType) {
-      case "image/jpeg":
-        return Uint8List.fromList(img.encodeJpg(image));
-      case "image/png":
-        return Uint8List.fromList(img.encodePng(image));
-      default:
-        throw UnimplementedError("Only Jpeg and Png are supported.");
-    }
+  Future<Uint8List> get bytes async {
+    Uint8List object = await file.readAsBytes();
+    return object;
   }
 }